backend/requirements/all.txt
Django==3.2
cx-Oracle==8.1.0
httpx==0.17.1
sentry-sdk==1.0.0
PyJWT==2.0.1
supervisor==4.2.2
pymemcache==3.4.2
celery[redis,gevent]==5.0.5
django-celery-results==2.0.1
django-celery-beat==2.2.0
elasticsearch-dsl>=7.3.0,<8.0.0
gunicorn==20.1.0
